fitClub/requirements.txt
Django == 1.8.2
djangorestframework == 3.1.3
